Avoid imports from "providers"
