+++ providers/fab/tests/unit/fab/auth_manager/models/test_db.py
@@ -24,7 +24,7 @@ from alembic.autogenerate import compare_metadata
 from alembic.migration import MigrationContext
 from sqlalchemy import MetaData
 
-import providers.fab.src.airflow.providers.fab as provider_fab
+import airflow.providers.fab as provider_fab
 from airflow.settings import engine
 from airflow.utils.db import (
     compare_server_default,
