+++ providers/microsoft/azure/tests/unit/microsoft/azure/test_utils.py
@@ -17,10 +17,7 @@
 
 from __future__ import annotations
 
-import json
-import re
 from json import JSONDecodeError
-from os.path import dirname, join
 from typing import Any
 from unittest import mock
 from unittest.mock import MagicMock
@@ -237,27 +234,3 @@ def mock_response(status_code, content: Any = None, headers: dict | None = None)
     response.content = content
     response.json.side_effect = JSONDecodeError("", "", 0)
     return response
-
-
-def remove_license_header(content: str) -> str:
-    """Remove license header from the given content."""
-    # Define the pattern to match both block and single-line comments
-    pattern = r"(/\*.*?\*/)|(--.*?(\r?\n|\r))|(#.*?(\r?\n|\r))"
-
-    # Check if there is a license header at the beginning of the file
-    if re.match(pattern, content, flags=re.DOTALL):
-        # Use re.DOTALL to allow .* to match newline characters in block comments
-        return re.sub(pattern, "", content, flags=re.DOTALL).strip()
-    return content.strip()
-
-
-def load_json(*args: str):
-    with open(join(dirname(__file__), *args), encoding="utf-8") as file:
-        return json.load(file)
-
-
-def load_file(*args: str, mode="r", encoding="utf-8"):
-    with open(join(dirname(__file__), *args), mode=mode, encoding=encoding) as file:
-        if mode == "r":
-            return remove_license_header(file.read())
-        return file.read()
